fix: refactor merkle proof utilities in common
